Moved lexer tests to their own file
