+++ src/compiler/lexer/mod.rs
@@ -1,2 +1,4 @@
+mod tests;
+
 pub(crate) mod lexer;
 pub(crate) mod tokens;
